// src/features/Bills/types/Category.ts
export enum Category {
  HABITATION = 'HABITATION',
  TRANSPORT = 'TRANSPORT',
  FOOD = 'FOOD',
  SUBSCRIPTION = 'SUBSCRIPTION',
  PHONE_INTERNET = 'PHONE_INTERNET',
  HEALTH = 'HEALTH',
  EDUCATION = 'EDUCATION',
  LEISURE = 'LEISURE',
  PETS = 'PETS',
  CREDIT_CARD = 'CREDIT_CARD',
  UNEXPECTED = 'UNEXPECTED',
  OTHERS = 'OTHERS',
}